+++ Sources/Playbook/Components/Button/PBButton.swift
@@ -106,48 +106,6 @@ public enum PBIconPosition {
 struct PBButton_Previews: PreviewProvider {
   static var previews: some View {
     registerFonts()
-
-    return List {
-      Section("Button Variants") {
-        PBButtonStyle_Previews.previews
-      }
-
-      Section("Button Icon Positions") {
-        PBButton(
-          title: "Button with Icon on Left",
-          icon: PBIcon.fontAwesome(.user, size: .x1),
-          action: {}
-        )
-        PBButton(
-          title: "Button with Icon on Right",
-          icon: PBIcon.fontAwesome(.user, size: .x1),
-          iconPosition: .right,
-          action: {}
-        )
-      }
-      .listRowSeparator(.hidden)
-
-      Section("Circle Buttons") {
-        PBCircleStyle_Previews.previews
-      }
-
-      Section("Button Sizes") {
-        PBButton(
-          size: .small,
-          title: "Button sm",
-          action: {}
-        )
-        PBButton(
-          title: "Button md",
-          action: {}
-        )
-        PBButton(
-          size: .large,
-          title: "Button lg",
-          action: {}
-        )
-      }
-      .listRowSeparator(.hidden)
-    }
+    return ButtonsCatalog()
   }
 }
