+++ Sources/Playbook/Components/Image/PBImage.swift
@@ -72,31 +72,6 @@ public extension PBImage {
 
 public struct PBImage_Previews: PreviewProvider {
   public static var previews: some View {
-    List {
-      ForEach(PBImage.Size.allCases, id: \.rawValue) { size in
-        Section(size.name) {
-          HStack {
-            PBImage(
-              image: nil,
-              placeholder: Image("Forest", bundle: .module),
-              size: size,
-              cornerRadius: .rounded
-            )
-            Spacer()
-            PBImage(
-              image: Image("Forest", bundle: .module),
-              size: size,
-              cornerRadius: .sharp
-            )
-          }
-        }
-      }
-      .listRowBackground(Color.clear)
-
-      Section("No size") {
-        PBImage(image: Image("Forest", bundle: .module))
-      }
-      .listRowBackground(Color.clear)
-    }
+   ImageCatalog()
   }
 }
