+++ Sources/Playbook/Components/Progress Indicator/PBSpinner.swift
@@ -58,6 +58,5 @@ public struct PBSpinner: View {
 public struct PBSpinner_Previews: PreviewProvider {
   public static var previews: some View {
     PBSpinner()
-      .padding()
   }
 }
