+++ Sources/Playbook/Design Elements/Colors/ColorsCatalog.swift
@@ -8,6 +8,9 @@
 import SwiftUI
 
 public struct ColorsCatalog: View {
+
+  public init() {}
+
   public var body: some View {
     let shape = Circle().frame(width: 60).pbShadow(.deep)
     List {
@@ -44,16 +47,16 @@ public struct ColorsCatalog: View {
         }
       }
 
-      Section("Status Subtle") {
-        LazyVGrid(columns: Array(repeating: GridItem(.flexible()), count: 4)) {
-          ForEach(Color.StatusColor.allCases, id: \.self) { color in
-            VStack {
-              shape.foregroundColor(Color.status(color, subtle: true))
-              Text(color.rawValue).pbFont(.subcaption, color: .text(.light))
-            }
-          }
-        }
-      }
+//      Section("Status Subtle") {
+//        LazyVGrid(columns: Array(repeating: GridItem(.flexible()), count: 4)) {
+//          ForEach(Color.StatusColor.allCases, id: \.self) { color in
+//            VStack {
+//              shape.foregroundColor(Color.status(color, subtle: true))
+//              Text(color.rawValue).pbFont(.subcaption, color: .text(.light))
+//            }
+//          }
+//        }
+//      }
 
 //      Section("Product Background") {
 //        LazyVGrid(columns: Array(repeating: GridItem(.flexible()), count: 4)) {
@@ -88,11 +91,5 @@ public struct ColorsCatalog: View {
 //        }
 //      }
     }
-  }
-}
-
-public struct ColorsCatalog_Previews: PreviewProvider {
-  public static var previews: some View {
-    ColorsCatalog()
   }
 }
